refactor(commands): extract helpInfo in a new file
