+++ index.js
@@ -1,5 +1,18 @@
 import { bot } from "./src/initConfig.js";
 
+import { smileys } from "./src/constantes/smileys.js";
+import { chanIds } from "./src/constantes/chanIds.js";
+import { randInt } from "./src/utils/randInt.js";
+
+import {
+  commandList,
+  pseudoCommandList,
+  invisibleCommandList,
+} from "./src/commandes/commandList.js";
+import { basicCommands } from "./src/commandes/basicCommands.js";
+import { pseudoCommands } from "./src/commandes/pseudoCommands.js";
+import { helpInfo } from "./src/commandes/helpInfo.js";
+
 import {
   newMessage,
   newCommand,
@@ -11,66 +24,10 @@ import {
   checkAnniversaire,
 } from "./profil.js";
 
-import { randInt } from "./src/utils/randInt.js";
-import { smileys } from "./src/constantes/smileys.js";
-import { chanIds } from "./src/constantes/chanIds.js";
-import {
-  commandList,
-  pseudoCommandList,
-  invisibleCommandList,
-} from "./src/commandes/commandList.js";
-import { basicCommands } from "./src/commandes/basicCommands.js";
-import { pseudoCommands } from "./src/commandes/pseudoCommands.js";
-
 // Variables globales
 const hll = new Date("2022", "07", "09", "20", "08", "00");
 let generalChan; // Objet contenant les réfs vers le canal #general
 
-const helpInfo = {
-  help: "**Utilisation**\n```!help _Commande```\n**Description**\n```La commande help donne des informations sur l'utilisation de la commande spécifiée. Si aucune commande n'est spécifiée, elle donne la liste des commandes disponibles.```",
-  membres:
-    '**Utilisation**\n```!membres```\n**Description**\n```La commande membres donne la liste des commandes disponibles sur certains membres ou ex membres. Ces commandes permettront de connaître des phrases drôles ou typiques prononcées par ces membres.```\n**Exemple**\n```!WassaW pourra donner "AHAHAHAHAH".```',
-  forum:
-    "**Utilisation**\n```!forum```\n**Description**\n```J'espère pour toi que tu sais à quoi sert le forum !```",
-  trombi:
-    "**Utilisation**\n```!trombi```\n**Description**\n```Le trombi te permet de voir la tête des autres si tu es prêt à montrer la tienne !```",
-  everyone:
-    "**Utilisation**\n```!everyone```\n**Description**\n```N'utilisez pas cette commande !!!```",
-  echo: "**Utilisation**\n```!echo```\n**Description**\n```Une commande tellement puissante que Cleme a dû en limiter l'effet...```",
-  cui: "**Utilisation**\n```!cui```\n**Description**\n```Un pile ou face créé par MacKay. Il y a donc deux réponses possibles : cui ou...```",
-  nuclear:
-    "**Utilisation**\n```!nuclear```\n**Description**\n```Le saviez-vous ? Bernard s'est fait hacké et s'est fait contrôler par une méchante personne dans ses débuts. Maintenant sa protection a été revue à la hausse mais cette commande est là pour rappeler cet évennement plutôt spécial...```",
-  cyanure:
-    "**Utilisation**\n```!cyanure```\n**Description**\n```A utiliser quand les gens spamment trop sur un canal.\n~ Commande créée par un poussin qui n'aime pas quand il y a trop de bruit```",
-  bleus:
-    "**Utilisation**\n```!bleus```\n**Description**\n```Les plus grands suporters de l'équipe de France sont sur ce discord !```",
-  actualité:
-    "**Utilisation**\n```!actualité```\n**Description**\n```Aussi à jour qu'internet explorer !```",
-  log: "**Utilisation**\n```!log```\n**Description**\n```Pour connaître les dernières nouveautés de Bernard.```",
-  logs: "**Utilisation**\n```!logs```\n**Description**\n```Historique des mises à jour de Bernard.```",
-  penis:
-    "**Utilisation**\n```!penis```\n**Description**\n```Qui réussira à battre Cleme ?```",
-  oreoplz:
-    "**Utilisation**\n```!oreoplz```\n**Description**\n```Je préfère ne pas en parler...```",
-  parapluie:
-    "**Utilisation**\n```!parapluie```\n**Description**\n```Demandez à Harty !```",
-  license:
-    "**Utilisation**\n```!license```\n**Description**\n```Si vous voulez copier Bernard, citez l'original !```",
-  d: '**Utilisation**\n```!d_nbPositif```\n**Description**\n```Donne un nombre aléatoire entre 1 et ce nombre.```\n**Exemple**\n```!d100 pourra donner "69".```',
-  peage:
-    "**Utilisation**\n```Indisponible```\n**Description**\n```Un jour cette commande existera !```",
-  profil:
-    "**Utilisation**\n```!profil _nomDiscord```\n**Description**\n```Affiche le profil de la personne donnée. Si aucune personne n'a été donnée, affiche votre ```",
-  rich: "**Utilisation**\n```!rich```\n**Description**\n```Donne la liste des personnes ayant le plus de Libcoins sur le discord. Mais à quoi peut bien servir cette monnaie ?```",
-  commandes:
-    "**Utilisation**\n```!commandes```\n**Description**\n```Donne la liste des commandes les plus utilisées sur ce discord.```",
-  classement:
-    "**Utilisation**\n```!classement```\n**Description**\n```Donne la liste des personnes ayant le plus parlé sur ce discord. Qui arrivera à détrôner Cleme ?```",
-  requete:
-    "**Utilisation**\n```!requete Nom Phrase```\n**Description**\n```Si vous voulez qu'une phrase dite par un joueur lui soit associée sur Bernard, vous pouvez utiliser cette commande. Si la communauté approuve, ces phrases seront ajoutées à Bernard. Comme ça les personnes qui ne sont pas dans les phrases de Bernard n'auront plus de raison de se plaindre !!```\n**Exemple**\n```!requete Khayrisill Tu me manques wawa```",
-  site: "**Utilisation**\n```!site```\n**Description**\n```Envoie un lien vers l'antre de Bernard. Qui sait ce que vous pourriez y trouver ?```",
-  film: "**Utilisation**\n```!film```\n**Description**\n```Envoie une citation aléatoire de l'un des films vu dans #cinema```",
-};
 let mat100 = 0;
 
 bot.on("ready", () => {
